b2b-automation-feature-activation/requirements.txt
Appium-Python-Client
behave
browserstack-local
ipdb
selenium
page_objects
paver
psutil
requests